+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/
 - Experimental: allow back-end selection through job option
 - Add support for `load_result` with a URL (instead of job_id).
   ([#95](https://github.com/Open-EO/openeo-aggregator/issues/95))
+- Batch job result metadata: preserve upstream's canonical link
+  ([#98](https://github.com/Open-EO/openeo-aggregator/issues/98))
 
 ### Changed